them phan don vi tinh cho trang pos

+++ src/modules/pos/components/cart/ScanModal.jsx
@@ -96,7 +96,7 @@ const ScanModal = ({ isOpen, onClose, onItemAdded, products = [] }) => {
               type="number"
               min="1"
               value={quantity}
-              onChange={(e) => setQuantity(parseInt(e.target.value) || 1)}
+              onChange={(e) => setQuantity(parseFloat(e.target.value) || 1)}
             />
           </div>
           <Button
